requirements.txt
matplotlib
pandas
grpcio
grpcio-tools